I added a reader for ddcMD files. It takes the extension ddcMD, ddcmd, and DDCMD.  The data from a ddcMD dump file consists of 1 or more files in a directory, so there isn't really a file to open, rather a directory. Since VisIt wants to open files, the work around it to create a zero length file with the same name as the directory with one of the supported extensions added to it.
It has both a serial and parallel implementation.  The parallel implementation
will automatically domain decompose the data.

The mesh for a ddcMD file is a regular 2d or 3d grid.  Unfortunately the data
is organized such that the data for a zone is stored contiguously in a record
and the records are arranged randomly throughout the file with an id at the
beginning of each record indicating which zone the data applies to.  This means
that all the data must be read in when the file is opened and a massive all
to all communication must be made to reorganize the data into rectilinear
blocks.



git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@2171 18c085ea-50e0-402c-830e-de6fd14e8384