User arguments for X server initialization.
This adds an `-x-args' command line option to the engine which allows the user
specify additional options when starting the X server.  The special formatter
`%l' is replaced with the display ID, and the special formatter `%n' is
replaced with the node ID.  It is the user's responsibility to make sure that
multiple arguments are properly quoted (i.e., they must appear as one argument
to the engine).


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@5099 18c085ea-50e0-402c-830e-de6fd14e8384