Add patches to build_visit so that BoxLib works.
This adds two patches to the BoxLib build for build_visit.  The
firts forces -fPIC; this is needed on x86_64 Linux systems, and
shouldn't hurt on anything else.  The second changes the compiler
from `g77' to `gfortran', and is only applied if the system does
not appear to have `g77' (no check is made for `gfortran'...).

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@5278 18c085ea-50e0-402c-830e-de6fd14e8384