Don't try to pass empty histograms to the widgets.
In the case that the plot hasn't been executed yet, we get back an
empty vector.  Thus operator[] with an argument of 0 (or any
argument, really) is out of bounds.

I believe this was only visible with debug iterators; the STL
implementations I use tend to ignore the case in release mode.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6472 18c085ea-50e0-402c-830e-de6fd14e8384