Define the HDF4 endianness correctly on linux-ppc systems.
This is a modification of a patch from Patrick Shinpaugh.  It
changes an internal HDF4 define from LITTLE_ENDIAN to BIG_ENDIAN
when compiling on linux-ppc.

Supposedly the compilation succeeds by just removing the changed
line, but that felt unsafe to me.  Patrick assures me that
compilation works without the define and with it defined as
BIG_ENDIAN.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6560 18c085ea-50e0-402c-830e-de6fd14e8384