Cripple Tuvok IO subsystem to avoid a Qt dependence.
Without this define, Tuvok needs Qt `imageformat' plugins, which
are painfully annoying to get setup correctly because they don't
appear in the standard library directories.  Tuvok is only using
those plugins to load and save images, and we do that via other
means in VisIt, so they're not needed anyway.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6601 18c085ea-50e0-402c-830e-de6fd14e8384