Rework enabling code so that the command line option takes precedence.
Previously the config-site var could override the configure
option, because of the strange order of sourcing / checking
AC_ARG_ENABLE variable setting.  This changes the configure-y
to use a new a variable instead of trying to re-use the enable
variable; this allows us to easily check whether the enable
variable was set, telling us that the user specified the value
manually.

As a strange side effect, AC_MSG_WARN seems to work now.  For some
reason, it was previously behaving like AC_MSG_ERROR (even on more
recent versions of autoconf), but after this patch, it (correctly)
warns + continues.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6751 18c085ea-50e0-402c-830e-de6fd14e8384