Set HW/SW rendering for the callback.
Outside of the engine, we don't know how many displays we'll have,
and which ones are hardware contexts.  External code (in
particular the volume plot) would use this callback object to
figure out what kind of renderers to create.

So we ignore the existing configuration and set hw/sw rendering
based on what we've figured out during initialization.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@7997 18c085ea-50e0-402c-830e-de6fd14e8384