Remove unneeded mode-specific baselines.
A lot of these actually looked like duplicates: the SR image was
the same as the serial image.  A minority actually pass now but
were failing due to an SR-specific baseline.

With the recent IceT and Mesa changes, it seems like it will be
less work to be overzealous here and re-add mode specific
baselines as needed.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@8691 18c085ea-50e0-402c-830e-de6fd14e8384