Add fallback OpenGL libraries for Linux, AIX.
Use the new colon-separated syntax to give a search order to the
system OpenGL shared object on Linux and AIX.  On Linux, we use
libGL.so.1, which should exist on all Linux systems, and fall back
to libGL.so, which should only exist on a developer's system,
should the runtime lib fail.  On AIX, one needs to encode the
machine's bit width into the library loading, because of how
system shared objects are built.  Note that our own shared
libraries are built a bit more sanely, and do not require the
"(object.o)" syntax, so the libOSMesa setting actually *is*
correct.

Also, the specifications for these libraries given all of our
platforms was getting a bit hairy, so I moved them into an ifdef
block above the struct for legibility.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@9211 18c085ea-50e0-402c-830e-de6fd14e8384