Fixed problem causing segmentation violation due to cut-n-paste error where arbMeshCellReMap entries were not getting cleared correctly if no new points were added to a potentially arbitrary polyehdral mesh.
Changed Silo's error level to DB_TOP instead of DB_ALL. That is because
newer versions of the Silo library do NOT call H5Eset_auto(0,0) for
DB_ALL and results in HDF5 error messages getting spewed to stderr.

Added siloDriver data member and setting of it in constructor and OpenFile.


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@10621 18c085ea-50e0-402c-830e-de6fd14e8384