Two different fixes.
1) Streamline code: src/avt/Pipeline/AbstractFilters/avtDatasetOnDemandFilter.C

If the dataset was not found the code would crash. I know check if the dataset is NULL and return NULL. The code will then throw an exception, but it will not crash the engine.

2) common/misc/VisItInit.C

Changed the order of the test in the for loop, so we don't look outside the memory of path. Before the change, the code would look at path[ -1 ].

     int  len = strlen(path);
     int lastSlash;
-    for (lastSlash=len ; path[lastSlash]!=visitcommon::SlashChar() && lastSlash>=0 ; lastSlash--)
+    for (lastSlash=len ; lastSlash>=0 && path[lastSlash]!=visitcommon::SlashChar() ; lastSlash--)
     {
         continue;
     }



git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@12725 18c085ea-50e0-402c-830e-de6fd14e8384