Do not force a particular compiler, but warn if it differs.
We attempted to force the compiler for out-of-tree plugins to be
the same one that VisIt was compiled with.  On systems with
compiler wrappers, i.e. "CC" that really either invoked or was an
alias for a g++ in some strange /opt location or something, this
would store the target compiler instead of the alias.  The wrapper
typically added flags which were then no longer propagated.

Furthermore, the existing logic was overly strict on compiler
versions.  If a sysadmin built VisIt with gcc 4.1.1 and then
installed gcc 4.1.2 somewhere else, our build system would force
using 4.1.1 even though the new compiler would be fine.

So instead of forcing a compiler, let's just warn that the
versions differ.  If the version difference matters, users will
get some notification; and if not, they can at least just ignore
the message without having to hack our .cmake files.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/branches/2.2RC@14061 18c085ea-50e0-402c-830e-de6fd14e8384