Support display strings that include screen types.
Old code assumed DISPLAYs of the form ":d", but many systems
utilize ":d.s" nowadays, where 'd' is the display number and 's'
is the screen number.  This commits rewrites the machinery to
store a string instead of a number for the display, and thus rely
on the user giving us the full DISPLAY string, allowing them to
set it to whatever the hell they feel like.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@14776 18c085ea-50e0-402c-830e-de6fd14e8384