I modified the AMRTest database reader to output a Mandelbrot image as its variable instead of a distance field. This is more interesting than the distance field. I also added information about the AMRTest database reader to the release notes under the "Changes for VisIt developers" section since I hadn't done that earlier.
Here is the information added to the release notes:

An AMRTest reader was added that demonstrates how to create a simple 
AMR database reader. It is only built when -DVISIT_BUILD_ALL_PLUGINS="1" 
is specified with cmake. It generates the data internally. To use the 
reader you should create a zero length file that ends with ".amrtest" 
and open it. It has 6 levels and has a single variable that creates a 
Mandelbrot image. It also works with the 2d multi resolution capabilities. 
To invoke them bring up the "Rendering options" window and select "Multi 
resolution for 2d AMR data" on the "Basic" tab.



git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@22442 18c085ea-50e0-402c-830e-de6fd14e8384