Build changes for LLNL RZ linux systems. (#88)
I rebuilt the third party libraries for 3.0.0b on the LLNL rztopaz system. I updated the rztopaz config site file and made the rzalaster and rztrona config site files links to the rztopaz one since the machines are the same.

I updated the LLNL open build and install scripts for building on LLNL RZ systems. I updated the rztrona build and install and removed the rzalastor one since a separate rzalastor was no longer necessary since rzalastor was now running the same operating system as rztopaz and rztrona.