Corrected a crash with compositing transparent geometry on cooley.alcf.anl.gov. (#92)
I corrected a bug that caused a crash compositing transparent geometry on cooley.alcf.anl.gov. The crash occurred whenever any memory that was allocated by aligned_alloc was freed. I made the use of aligned_alloc dependent on USE_ALIGNED_ALLOC being defined, which it won't be unless someone sets it as a preprocessor flag.