Fix crash on startup on Windows when system opengl version too old. (#4276)
* Fix crash on startup on Windows when system opengl version too old.

Changes made in visit-deps #18 to suppor this: add Mesa3D's opengl32.dll. Update installer to test OpenGL version and drop in the Mesa3D version if necessary. Add RegKey if necessary.

Modified launcher on Windows to check for VISITNEEDSMESA regkey and add 'MESA_GL_VERISON_OVERRIDE=3.3' to environment for viewer and engine processes if this key is set.

Update windows config-site file to set VISIT_MESA_REPLACE_OPENGL and VISIT_MESA_GL_DIR.

Modified Find modules for MesaGL and OpenGL to handle the Windows case when VISIT_MESA_REPLACE_OPENGL is set to true.

Modifed Find module for VTK to look for vtkTestOpenGLVerison.exe and install it.

Removed InitVTKRenderingConfig.h.in, as its only purpose was to set VISIT_MANTA flag, and Manta is no longer used.