Updates stock host profile for rhea.ccs.ornl.gov (#4689)
ORNL's Rhea cluster migrated to using slurm for scheduling over a year ago. And the location of the facility-installed VisIT binaries changed at the same time. This commit updates the stock host profile for Rhea to use sbatch/srun as the default launch method and sets the default path to VisIT to the new, correct location.

Since it's not possible to start jobs without specifying an project ID for allocation accounting on Rhea, this commit also sets the default `bankSet` boolean checkbox to `true` and adds a suggestive placeholder value for the default project ID (`bank` field).