Fix display of -help info on Windows
Modified to use visitusage.txt and to alloc a new console if Visit was bult as a Windows app. (MessageBox doesn't work well for large amounts of text.) This resolves #5088.

Added retrieval of git revision hash for use with -git-revision.
Update visitusage.txt to reflect that the correct command line is 'git-revision' not 'git-version'.
Update VisItInit to use a MessageBox on Windows for display of -git-revision.