Fix auto-color selection modes for Curves, Meshes and Lineouts
For cases where VisIt selects a color for a plot (Curve, Mesh and Lineout), the selection logic was a) fixed to avoid having it change (#5607) and b) improved to deconflict it with background color. In addition, we also ensure that VisIt's color selection is reflected in the GUI, where appropriate, even if that related GUI elements are not enabled at the time.