Add verbiage about expired SRE issues
I adjusted text here to deal with *expired* issues. I propose we introduce a new SRE tracker label, `expired`, when we close SRE issues for this reason. I don't know whether we need the same `expired` label in our PD tracker. I don't think so.

I was going to include verbiage indicating VisIt developers should make 2 or 3 attempts to re-engage the user prior to expiring. However, I opted not to do that because I think the *true* measure for how important the issue is to the user is whether *they* continue to engage us (e.g. bat the ball back) without our prodding. 

That assumes we are operating *always* with the thinking cap of finding some way to be helpful (e.g. hang on to the ball as much as possible and only bat it back to the user when necessary) as opposed to finding some reason, any reason to put the onis back on the user (e.g. continue batting the ball back to them). I think we do a good job of that and so don't think it is ok to be vague on this point in our SRE policies.