Fix importing VisIt into python on Windows
Add -nodialog option (unadvertised) for use with 'visit -env' on Windows to ensure that environment string is retrievable and that no dialog message box is created.
Add LIBPATH to environment on WINDOWS.
For frontend.py, if 'vdir' isn't set in a LaunchCommand, see if 'VISITLOC' is set in environment, and use it for vdir.