Change order of fields for Contour Plot.
Size of the MultiColor ColorAttributeList depends on the value of contourNLevels, contourPercent or contourValue. When any of those are changed,  'EnlargeMultiColor' is called, ensuring MultiColor has the right size.

The order of setting these fields in python is important, if SetMultiColor is called before the other fields are changed, it can cause an indexing error if the MultiColor object isn't the right size. This can be seen by using CommandRecording and changing default contourNLevels from 10 to 15, then attempt to execute what was recorded.

Reordering the fields allows them to executed in the correct order to ensure MultiColor has correct size before being modified.