Add CMakeUserPresets to .gitignore
Newer CMake has the concept of 'presets' that can be used to pre-fill cache items or setup a build. Global presets are stored in the root directory and are named 'CMakePresets.json'. User presets are also stored in the root directory and are named 'CMakeUserPresets.json'.  User presets shouldn't be added to repo and tracked, so I added them to .gitignore.