Update CMake version to 3.24.2. (#18272)
* Update CMake version to 3.24.2.
Remove flags that turned off CMAKE_USE_OPENSSL. Resolves #17927.
Add openssl-devel to fedora31 dockerfile so that CMake will build there.

* Fix ci for new cmake.

* Add '--recursive' to git clone command.

* Fix ci cmake-configure error.

* Add CMakeUserPresets to .gitignore

Newer CMake has the concept of 'presets' that can be used to pre-fill cache items or setup a build. Global presets are stored in the root directory and are named 'CMakePresets.json'. User presets are also stored in the root directory and are named 'CMakeUserPresets.json'.  User presets shouldn't be added to repo and tracked, so I added them to .gitignore.