Rm bizarre last line (#18348)
The last line had no effect because the `exec` statements in both branches of the `if-else` block meant it was never executed. On top of that, there is no clear idea what that line was *intended* to do. What it did do was call `frontendlauncher` (`$0`) recursively with three args, `=`, `shift` and `@ARGV;` which I think was not intentional.