Fix spatial extents retrieval for vtkPolyData with VTK-9.
Previous behavior of 'GetBounds' call on poly data was to return the extents of the cells, new behavior is to return the extents of the point set, which may not be fully connected to cells.
To preserve previous behavior, call GetCellsBounds with vtk 9.

This showed up in the  reflect operator test, where Boundary plot of Globe was projected and reflected.

Also removed an unnecessary call to GetBounds in the Neighbor expression because the results were never used.