Fix bug with zonecenter query.
Evidenced with more-procs-than-domains situation, and data that has ghost zones.
Exacerbated by the fact that logic to have only the proc that owns the queried domain perform the query doesn't work.

If a proc that never had a domain participated in the query, when it read in the mesh, it didn't have the right ghost information and would give different results based on that.

So the quick fix is to prevent any proc that never had a domain from participating in the query.

Added the same fix to  the Node Coords query.