Allow multiple domains to pass through the renderer.
rendering/volume.py test `volume_20` tests a multi-domain rectilinear grid with `resampling` turned off.  The results are a single-domain, because the renderer for the plot expects there to be only one input.
I modified the renderer to allow multiple inputs.
The results now show all domains, but because they are all rendered separately the boundaries between the domains are shown.
I think this is okay, because the user is presented with a warning that resampling is recommended and the results may not look good with out it.