Fix literal include that referenced a file outside the 'src' dir. (#19386)
This was a problem when building the manuals target from a released source tarball, because the source tarball does not have the `scripts` dir.
I copied the referenced file and added it to a `.. code:: yaml` directive.
I also removed references to frontier's `compilers.yaml` file as it no longer exists.
This resolves #19379.