Fix qt6 build issues when qt6 on system (#19618)
* Modified the builds of qttools and qtsvg to use qt-configure-module for configuring instead of CMake.
There were issues building Qt 6 if there was a version installed on the sytstem.
qttools would pick up some of the qtbase modules from the system instead of using the ones pointed to by Qt6_DIR passed to CMake.

* Add use of Qt6CoreTools_DIR, Qt6GuiTools_DIR, and Qt6WidgetsTools_DIR
to ensure that VTK utilizes only the Qt6 modules from the Qt6 that build_visit points to and not mix with other versions that might be on the system.

* Update release notes with recent fixes to build_visit.