Update regression test to build/run outside of the repo directory.
Added repoDir and buildDir vars to facilitate changing directories between the build dir and the repo dir.

Added call to 'git submodule init' when cloning the repo.
Added call to 'git submodule update' when updating the repo.
Removed call to 'git lfs pull' in test and data dirs, since the command is run from the root of the repo checkout, and shouldn't need to be called again for subdirs.
Create the build dir outside of the repo, and run the tests from that dir, to prevent pollution of the repo with build/test artifacts.