Make qwt optional and allow use of system Qwt (#19680)
* Make QWT optional.

Add HAVE_QWT definition to gui when QWT available.
Make StripChart classes only build when QWT available.
Update FindQWT.cmake so that it is not an error if QWT not found or requested.

* Make building of Qwt optional.
Update bv_qwt to support --system-qwt and fix --alt-qwt-dir.
Add --qwt to the build_visit invocations of run-build-visit.
Don't add StripChart gui classes to the build unless QWT found.

* Add @markcmiller86's suggestion of a better comment.

Co-authored-by: Mark C. Miller <miller86@llnl.gov>

---------

Co-authored-by: Mark C. Miller <miller86@llnl.gov>