Add support for GEOS flavor .pvd root files and .vtm timestate files.
Each .vtm file is considered as one mesh.
Blocks with 'Region' in the name that are Parent to 'Dataset' Blocks are considered groups.
Variables defined in the underlying .vtu files are tracked for which blocks they belong to, since they may not be defined in every group.
GetMesh and GetVar calls check the validity of the mesh/variable for a given block and  return nullptr when not considered valid.