Update View2DAttributes, View3DAttributes,DabaseCorrelation for python
Added all custom code to their .code files with comments specifying the customizations for ease when comparisons after future regeneration.

Updated GeneratePython to allow for .code file specialized _Notify, _dir, _methods, as well as SLOT.

Regenerated the Py files for these.

There were a lot of changes due to the fact that they have been on the 'do not regenerate list' for quite awhile.