Fix issues with condition defines in .code files on Windows.
The correct code to write to the CMakeLists.txt is 'add_compile_definitions' instead of 'add_definitions'.  The latter requires '-D' to be the prefix to whatever definitions are being added.
Added support for conditional cxx flags to handle an instance of '-f' flag. Changed the .code file to use 'CXXFlags' instead of 'Definitions', and added code to xmlcmake to write 'add_compile_options'.
This fixes compile failures on Windows.